+++ src/matrices/DenseSerialMatrix.h
@@ -79,7 +79,7 @@ public:
       * on the actual subclass of matrix used.
       */
     virtual void
-    addValuesByGlobalID( int num_rows, int num_cols, int *rows, int *cols, double *values );
+    addValuesByGlobalID( size_t num_rows, size_t num_cols, size_t *rows, size_t *cols, double *values );
 
     /** \brief  Set values in the matrix
       * \param[in] num_rows The number of rows represented in values
@@ -92,7 +92,7 @@ public:
       * on the actual subclass of matrix used.
       */
     virtual void
-    setValuesByGlobalID( int num_rows, int num_cols, int *rows, int *cols, double *values );
+    setValuesByGlobalID( size_t num_rows, size_t num_cols, size_t *rows, size_t *cols, double *values );
 
     /** \brief  Get values in the matrix
       * \param[in] num_rows The number of rows represented in values
@@ -104,7 +104,7 @@ public:
       *   have not been allocated or are not ghosts on the current processor.
       */
     virtual void
-    getValuesByGlobalID( int num_rows, int num_cols, int *rows, int *cols, double *values ) const;
+    getValuesByGlobalID( size_t num_rows, size_t num_cols, size_t *rows, size_t *cols, double *values ) const;
 
 
     /** \brief  Add values to those in the matrix
@@ -115,7 +115,7 @@ public:
       * allocated a particular(row,col) specified, depending
       * on the actual subclass of matrix used.
       */
-    virtual void addValueByGlobalID( int row, int col, double value );
+    virtual void addValueByGlobalID( size_t row, size_t col, double value );
 
     /** \brief  Set values in the matrix
       * \param[in] row  The row id of value
@@ -125,7 +125,7 @@ public:
       * allocated a particular(row,col) specified, depending
       * on the actual subclass of matrix used.
       */
-    virtual void setValueByGlobalID( int row, int col, double value );
+    virtual void setValueByGlobalID( size_t row, size_t col, double value );
 
     /** \brief  Set values in the matrix
       * \param[in] row  The row id of value
@@ -134,7 +134,7 @@ public:
       * allocated a particular(row,col) specified, depending
       * on the actual subclass of matrix used.
       */
-    virtual double getValueByGlobalID( int row, int col ) const;
+    virtual double getValueByGlobalID( size_t row, size_t col ) const;
 
 
     /** \brief  Set the non-zeros of the matrix to a scalar
@@ -149,7 +149,7 @@ public:
       * \param[out] values  The values in the row
       */
     virtual void
-    getRowByGlobalID( int row, std::vector<unsigned int> &cols, std::vector<double> &values ) const;
+    getRowByGlobalID( size_t row, std::vector<size_t> &cols, std::vector<double> &values ) const;
 
     /** \brief  Set the diagonal to the values in a vector
       * \param[in] in The values to set the diagonal to
