+++ src/matrices/petsc/NativePetscMatrix.h
@@ -64,13 +64,13 @@ public:
     virtual void axpy( double alpha, const Matrix &x ) override;
 
     virtual void
-    addValuesByGlobalID( int num_rows, int num_cols, int *rows, int *cols, double *values ) override;
+    addValuesByGlobalID( size_t num_rows, size_t num_cols, size_t *rows, size_t *cols, double *values ) override;
     virtual void
-    setValuesByGlobalID( int num_rows, int num_cols, int *rows, int *cols, double *values ) override;
+    setValuesByGlobalID( size_t num_rows, size_t num_cols, size_t *rows, size_t *cols, double *values ) override;
     virtual void
-    getValuesByGlobalID( int num_rows, int num_cols, int *rows, int *cols, double *values ) const override;
+    getValuesByGlobalID( size_t num_rows, size_t num_cols, size_t *rows, size_t *cols, double *values ) const override;
     virtual void
-    getRowByGlobalID( int row, std::vector<unsigned int> &cols, std::vector<double> &values ) const override;
+    getRowByGlobalID( size_t row, std::vector<size_t> &cols, std::vector<double> &values ) const override;
 
     virtual void setScalar( double ) override;
     virtual void setDiagonal( Vector::const_shared_ptr in ) override;
